+++ src/Registry/DoctrineTenantRegistry.php
@@ -49,6 +49,14 @@ final readonly class DoctrineTenantRegistry implements TenantRegistryInterface
         return $tenant instanceof TenantInterface ? $tenant : null;
     }
 
+    public function findById(string|int $id): ?TenantInterface
+    {
+        $repository = $this->em->getRepository($this->tenantEntityClass);
+        $tenant = $repository->find($id);
+
+        return $tenant instanceof TenantInterface ? $tenant : null;
+    }
+
     public function hasSlug(string $slug): bool
     {
         return null !== $this->findBySlug($slug);